requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
anthropic==0.52.0
python-dotenv==1.1.0
gspread==6.2.0
google-auth==2.40.1
pydantic==2.11.3
